uv-supabase-demo/pyproject.toml
[project]
name = "uv-supabase-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.6",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "supabase>=2.18.1",
]
